Refactor graphql package and class names
